Remove exception_handler module
