+++ mreg_cli/utilities/api.py
@@ -160,7 +160,7 @@ def prompt_for_password_and_login(user: str, url: str, catch_exception: bool = T
         auth_and_update_token(user, password)
     except CliError as e:
         if catch_exception:
-            from mreg_cli.exception_handler import handle_exception  # noqa: PLC0415
+            from mreg_cli.exceptions import handle_exception  # noqa: PLC0415
 
             handle_exception(e)
         if isinstance(e, LoginFailedError):
@@ -189,7 +189,7 @@ def prompt_for_password_and_try_update_token() -> None:
             raise LoginFailedError("Unable to determine username.")
         auth_and_update_token(user, password)
     except CliError as e:
-        from mreg_cli.exception_handler import handle_exception  # noqa: PLC0415
+        from mreg_cli.exceptions import handle_exception  # noqa: PLC0415
 
         handle_exception(e)
         raise e  # FIXME: remove if redundant
